[HUDI 1623] 1.Introducing start & end commit times to timeline filters             2. Instant time granularity changed to milliseconds             3. Added versioning to HoodieInstant disk format
Fix checkstyle

+++ hudi-cli/src/test/java/org/apache/hudi/cli/commands/TestTableCommand.java
@@ -115,7 +115,7 @@ public class TestTableCommand extends CLIFunctionalTestHarness {
 
     // Check default values
     assertFalse(conf.isConsistencyCheckEnabled());
-    assertEquals(new Integer(1), HoodieCLI.layoutVersion.getVersion());
+    assertEquals(new Integer(2), HoodieCLI.layoutVersion.getVersion());
   }
 
   /**
@@ -132,7 +132,7 @@ public class TestTableCommand extends CLIFunctionalTestHarness {
     assertEquals(tablePath, client.getBasePath());
     assertEquals(metaPath, client.getMetaPath());
     assertEquals(HoodieTableType.COPY_ON_WRITE, client.getTableType());
-    assertEquals(new Integer(1), client.getTimelineLayoutVersion().getVersion());
+    assertEquals(new Integer(2), client.getTimelineLayoutVersion().getVersion());
   }
 
   /**
